fix(codegen): DEPYLER-1135 Remove unnecessary numeric casts in push operations
The generate_typed_push function was unconditionally casting integers to i64
and floats to f64. This caused E0308 type mismatch errors when pushing to
Vec<i32> because Python's `int` maps to i32 by default, not i64.

Fix:
- Removed `as i64` cast for Type::Int - trust the type system
- Removed `as f64` cast for Type::Float - trust the type system
- If there's a type mismatch, it should be fixed at the source (type inference)
  not papered over with casts

Also fixes missing struct fields (module_aliases, alias) in test files that
were added in DEPYLER-1136-1139 but not updated in all test locations.

Results:
- Filtered corpus (non-test files): 32.6% → 37.2% (+2 files compiling)
- All 11,309 tests pass
- functional_programming_combined.py now compiles (was failing with E0308)

+++ examples/test_v02.rs
@@ -2616,7 +2616,7 @@ pub fn process_list(items: &Vec<i32>) -> Vec<i32> {
     let mut result = vec![];
     for item in items.iter().cloned() {
         if item > 0 {
-            result.push((item).py_mul(2) as i64);
+            result.push((item).py_mul(2));
         }
     }
     result
